+++ src/DevilDaggersInfo.Tools/NativeInterface/Services/OSX/OSXWindowingService.cs
@@ -1,0 +1,12 @@
+using System.Numerics;
+
+namespace DevilDaggersInfo.Tools.NativeInterface.Services.OSX;
+
+internal sealed class OSXWindowingService : INativeWindowingService
+{
+	public Vector2 GetWindowPosition()
+	{
+		// TODO: Implement.
+		return default;
+	}
+}